I have move the sentMail api route outer side and modified the mail style
